+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - [#28](https://github.com/Kashoo/synctos/issues/28): Parameter to allow unknown properties in a document or object
+- [#49](https://github.com/Kashoo/synctos/issues/49): Explicitly declare JSHint rules
 
 ## [1.2.0]
 ### Added